new_requirements.txt
# Core FastAPI
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# ML/Image Processing
tensorflow==2.13.0
numpy==1.24.3
opencv-python-headless==4.8.1
pillow==10.0.1
matplotlib==3.7.2

# Authentication (if you use JWT)
python-jose==3.3.0

# Utilities
python-dotenv==1.0.0